LayoutTests/fast/js/dfg-double-addition-simplify-to-int-expected.txt

Tests that the DFG doesn't get confused about an edge being a double edge after we perform CFG simplification.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(42, true)[0] is 84 on all iterations including after DFG tier-up.
PASS successfullyParsed is true

TEST COMPLETE
